Add Support for Member Info in Commands
[Problem]
Right now, commands are not able to see any
information about the member that issued
a command.

[Solution]
Added support for passing in the member's
info when commands are issued. Setup the
"hello" command to use this info.

[Testing]
Updated the hello command's unit test, and
also updated the command-runner to support
member info, testing both manually as well.